+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ## Changelog ##
 #### dev ####
 
+#### v2.5.0 ####
+* Fix `DamageModApplyingRelic` not affecting cards
+* Make `BindObjectToDamageInfo` patch public so it can be patched like it says
+* Powers
+  * `OnDrawPileShufflePower` (Vex)
+* Cards
+  * `OnObtainCard` (Vex)
+
 #### v2.4.2 ####
 * Fix Stun leaving base intent damage unchanged
   * Caused cards like Spot Weakness to sometimes incorrectly identify enemies as attacking when they're stunned